fixes, lighting
major changes to lighting, try to to fix fractional offset again *pulls hair"
other small fixes and changes.
lots of good changes in volumetrics, reflections are performant and working properly now (sides of buildings especially) no more temporal tricks are being used in light compute shader, which was an attempt to fix the fractional offset problem that remains only for lighting. the fractional offset must be added somewhere that I'm missing.
any kind of clear that can be done asynchronously with other work is now done in parallel safetly. lots of memsets and memcpy's optimized.