big update - stream compaction
parallel iteration is inherently unordered. so when using a really high amount of voxels, the voxel vertex buffer was glitched Now using stream compaction to send to the gpu for all voxels (terrain todo)

lighting has changed for the better. the resolution is limited to 128 cubed for performance.  results in light jumping. 256 fixes the problem but overall tanks performance.